+++ src/main/java/teamroots/embers/config/compat/EnvironmentalTechCategory.java
@@ -4,11 +4,44 @@ import net.minecraftforge.common.config.Config;
 
 public class EnvironmentalTechCategory {
     @Config.RequiresMcRestart
-    @Config.Name("Energy Increase")
-    @Config.RangeInt(min = 0)
+    @Config.Name("Ember Modifier Energy Increase")
     @Config.Comment({
-            "How much more energy the modifier will consume",
-            "The value is entered as a percentage, so 20 is 20% more energy"
+            "How much more energy the ember modifier will consume.",
+            "The value is treated as a percentage, so 20 increases the energy consumption by 20%."
     })
-    public int energy_increase = 20;
+    @Config.RangeInt(min = 0)
+    public int ember_energy_increase = 20;
+
+    @Config.RequiresMcRestart
+    @Config.Name("Luck Modifier Energy Increase")
+    @Config.Comment({
+            "How much more energy the luck modifier will consume.",
+            "The value is treated as a percentage, so 90 increases the energy consumption by 90%.",
+            "NOTE: This is an additional config option, because Environmental Tech",
+            "doesn't provide an option for this by default."
+    })
+    @Config.RangeInt(min = 0)
+    public int luck_energy_increase = 90;
+
+    @Config.RequiresMcRestart
+    @Config.Name("Speed Modifier Energy Increase")
+    @Config.Comment({
+            "How much more energy the speed modifier will consume.",
+            "The value is treated as a percentage, so 100 increases the energy consumption by 100%.",
+            "NOTE: This is an additional config option, because Environmental Tech",
+            "doesn't provide an option for this by default."
+    })
+    @Config.RangeInt(min = 0)
+    public int speed_energy_increase = 100;
+
+    @Config.RequiresMcRestart
+    @Config.Name("Accuracy Modifier Energy Increase")
+    @Config.Comment({
+            "How much more energy the accuracy modifier will consume.",
+            "The value is treated as a percentage, so 50 increases the energy consumption by 50%.",
+            "NOTE: This is an additional config option, because Environmental Tech",
+            "doesn't provide an option for this by default."
+    })
+    @Config.RangeInt(min = 0)
+    public int accuracy_energy_increase = 50;
 }